Feature: buffered strings (#1)
Top line: everything in here should work as expected except string->symbol can't do interning as we really want hash tables for that, which in turn really want object vectors.

* WIP: array-backed strings and u8vectors are turned on in the assembler and in primitives. everything runs, but some tests still fail.
* Debug + fix string->list, which fixed many (but not all) tests
* All tests pass, caught a few other subtle issues with -fsanitize flags
* WIP more extensive symbol support
* Small fixes to u8vectors and symbols after a bunch of debugging and learning how the GC works
* Fix sweeping of vector-backed strings, updated comments
* Better support for escape sequences and printing strings and vectors

+++ vm/core/dispatch.c
@@ -79,7 +79,7 @@ uint8 handle_arity_and_rest_param (uint8 na)
 			ERROR("handle_arity_and_rest_param.0", "wrong number of arguments");
 		}
 	} else {
-		np = ~np;
+		np = (uint8)(~np);
 
 		if (na < np) {
 			ERROR("handle_arity_and_rest_param.1", "wrong number of arguments");
